create catalog in BeforeSuite, clean up unnecessary timeout+polling
