+++ test/e2e/install_test.go
@@ -3,7 +3,6 @@ package e2e
 import (
 	"context"
 	"fmt"
-	"time"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
@@ -17,44 +16,15 @@ import (
 	"k8s.io/apimachinery/pkg/util/rand"
 )
 
-const (
-	defaultTimeout  = 30 * time.Second
-	defaultPoll     = 1 * time.Second
-	testCatalogRef  = "localhost/testdata/catalogs/test-catalog:e2e"
-	testCatalogName = "test-catalog"
-)
-
 var _ = Describe("Operator Install", func() {
 	var (
-		ctx             context.Context
-		operatorName    string
-		operator        *operatorv1alpha1.Operator
-		operatorCatalog *catalogd.Catalog
+		ctx          context.Context
+		operatorName string
+		operator     *operatorv1alpha1.Operator
 	)
 	When("An operator is installed from an operator catalog", func() {
 		BeforeEach(func() {
 			ctx = context.Background()
-			var err error
-			operatorCatalog, err = createTestCatalog(ctx, testCatalogName, testCatalogRef)
-			Expect(err).ToNot(HaveOccurred())
-
-			Eventually(func(g Gomega) {
-				err := c.Get(ctx, types.NamespacedName{Name: operatorCatalog.Name}, operatorCatalog)
-				g.Expect(err).ToNot(HaveOccurred())
-				cond := apimeta.FindStatusCondition(operatorCatalog.Status.Conditions, catalogd.TypeUnpacked)
-				g.Expect(cond).ToNot(BeNil())
-				g.Expect(cond.Status).To(Equal(metav1.ConditionTrue))
-				g.Expect(cond.Reason).To(Equal(catalogd.ReasonUnpackSuccessful))
-				// g.Expect(cond.Message).To(ContainSubstring("resolved to"))
-
-				// Ensure some packages exist before continuing so the
-				// operators don't get stuck in a bad state
-				pList := &catalogd.PackageList{}
-				err = c.List(ctx, pList)
-				g.Expect(err).ToNot(HaveOccurred())
-				g.Expect(pList.Items).To(HaveLen(2))
-			}).WithTimeout(defaultTimeout).WithPolling(defaultPoll).Should(Succeed())
-
 			operatorName = fmt.Sprintf("operator-%s", rand.String(8))
 			operator = &operatorv1alpha1.Operator{
 				ObjectMeta: metav1.ObjectMeta{
@@ -84,7 +54,7 @@ var _ = Describe("Operator Install", func() {
 					g.Expect(cond.Reason).To(Equal(operatorv1alpha1.ReasonSuccess))
 					g.Expect(cond.Message).To(ContainSubstring("resolved to"))
 					g.Expect(operator.Status.ResolvedBundleResource).ToNot(BeEmpty())
-				}).WithTimeout(defaultTimeout).WithPolling(defaultPoll).Should(Succeed())
+				}).Should(Succeed())
 
 				By("eventually installing the package successfully")
 				Eventually(func(g Gomega) {
@@ -102,7 +72,7 @@ var _ = Describe("Operator Install", func() {
 					g.Expect(len(bd.Status.Conditions)).To(Equal(2))
 					g.Expect(bd.Status.Conditions[0].Reason).To(Equal("UnpackSuccessful"))
 					g.Expect(bd.Status.Conditions[1].Reason).To(Equal("InstallationSucceeded"))
-				}).WithTimeout(defaultTimeout).WithPolling(defaultPoll).Should(Succeed())
+				}).Should(Succeed())
 			})
 		})
 
@@ -128,7 +98,7 @@ var _ = Describe("Operator Install", func() {
 					g.Expect(cond.Reason).To(Equal(operatorv1alpha1.ReasonSuccess))
 					g.Expect(cond.Message).To(ContainSubstring("resolved to"))
 					g.Expect(operator.Status.ResolvedBundleResource).ToNot(BeEmpty())
-				}).WithTimeout(defaultTimeout).WithPolling(defaultPoll).Should(Succeed())
+				}).Should(Succeed())
 
 				By("eventually installing the package successfully")
 				Eventually(func(g Gomega) {
@@ -146,7 +116,7 @@ var _ = Describe("Operator Install", func() {
 					g.Expect(len(bd.Status.Conditions)).To(Equal(2))
 					g.Expect(bd.Status.Conditions[0].Reason).To(Equal("UnpackSuccessful"))
 					g.Expect(bd.Status.Conditions[1].Reason).To(Equal("InstallationSucceeded"))
-				}).WithTimeout(defaultTimeout).WithPolling(defaultPoll).Should(Succeed())
+				}).Should(Succeed())
 			})
 		})
 
@@ -164,7 +134,7 @@ var _ = Describe("Operator Install", func() {
 				// target package should not be present on cluster
 				err := c.Get(ctx, types.NamespacedName{Name: pkgName}, &catalogd.Package{})
 				g.Expect(errors.IsNotFound(err)).To(BeTrue())
-			}).WithTimeout(5 * time.Minute).WithPolling(defaultPoll).Should(Succeed())
+			}).Should(Succeed())
 
 			By("creating the Operator resource")
 			err = c.Create(ctx, operator)
@@ -179,7 +149,7 @@ var _ = Describe("Operator Install", func() {
 				g.Expect(cond.Status).To(Equal(metav1.ConditionFalse))
 				g.Expect(cond.Reason).To(Equal(operatorv1alpha1.ReasonResolutionFailed))
 				g.Expect(cond.Message).To(Equal(fmt.Sprintf("package '%s' not found", pkgName)))
-			}).WithTimeout(defaultTimeout).WithPolling(defaultPoll).Should(Succeed())
+			}).Should(Succeed())
 
 			By("creating an Operator catalog with the desired package")
 			operatorCatalog, err = createTestCatalog(ctx, testCatalogName, testCatalogRef)
@@ -191,7 +161,7 @@ var _ = Describe("Operator Install", func() {
 				g.Expect(cond).ToNot(BeNil())
 				g.Expect(cond.Status).To(Equal(metav1.ConditionTrue))
 				g.Expect(cond.Reason).To(Equal(catalogd.ReasonUnpackSuccessful))
-			}).WithTimeout(5 * time.Minute).WithPolling(defaultPoll).Should(Succeed())
+			}).Should(Succeed())
 
 			By("eventually resolving the package successfully")
 			Eventually(func(g Gomega) {
@@ -201,67 +171,16 @@ var _ = Describe("Operator Install", func() {
 				g.Expect(cond).ToNot(BeNil())
 				g.Expect(cond.Status).To(Equal(metav1.ConditionTrue))
 				g.Expect(cond.Reason).To(Equal(operatorv1alpha1.ReasonSuccess))
-			}).WithTimeout(defaultTimeout).WithPolling(defaultPoll).Should(Succeed())
+			}).Should(Succeed())
 		})
 
 		AfterEach(func() {
 			err := c.Delete(ctx, operator)
 			Eventually(func(g Gomega) {
 				err = c.Get(ctx, types.NamespacedName{Name: operator.Name}, &operatorv1alpha1.Operator{})
-				Expect(errors.IsNotFound(err)).To(BeTrue())
-			}).WithTimeout(defaultTimeout).WithPolling(defaultPoll).Should(Succeed())
-
-			err = c.Delete(ctx, operatorCatalog)
-			Expect(err).ToNot(HaveOccurred())
-			Eventually(func(g Gomega) {
-				err = c.Get(ctx, types.NamespacedName{Name: operatorCatalog.Name}, &catalogd.Catalog{})
-				Expect(errors.IsNotFound(err)).To(BeTrue())
-			}).WithTimeout(defaultTimeout).WithPolling(defaultPoll).Should(Succeed())
-
-			// speed up delete without waiting for gc
-			err = c.DeleteAllOf(ctx, &catalogd.BundleMetadata{})
-			Expect(err).ToNot(HaveOccurred())
-			err = c.DeleteAllOf(ctx, &catalogd.Package{})
-			Expect(err).ToNot(HaveOccurred())
-
-			Eventually(func(g Gomega) {
-				// ensure resource cleanup
-				packages := &catalogd.PackageList{}
-				err = c.List(ctx, packages)
-				Expect(err).To(BeNil())
-				Expect(packages.Items).To(BeEmpty())
-
-				bmd := &catalogd.BundleMetadataList{}
-				err = c.List(ctx, bmd)
-				Expect(err).To(BeNil())
-				Expect(bmd.Items).To(BeEmpty())
-
-				err = c.Get(ctx, types.NamespacedName{Name: operatorName}, &rukpakv1alpha1.BundleDeployment{})
-				Expect(errors.IsNotFound(err)).To(BeTrue())
-			}).WithTimeout(5 * time.Minute).WithPolling(defaultPoll).Should(Succeed())
+				g.Expect(errors.IsNotFound(err)).To(BeTrue())
+			}).Should(Succeed())
 		})
 
 	})
 })
-
-// createTestCatalog will create a new catalog on the test cluster, provided
-// the context, catalog name, and the image reference. It returns the created catalog
-// or an error if any errors occurred while creating the catalog.
-func createTestCatalog(ctx context.Context, name string, imageRef string) (*catalogd.Catalog, error) {
-	catalog := &catalogd.Catalog{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: name,
-		},
-		Spec: catalogd.CatalogSpec{
-			Source: catalogd.CatalogSource{
-				Type: catalogd.SourceTypeImage,
-				Image: &catalogd.ImageSource{
-					Ref: imageRef,
-				},
-			},
-		},
-	}
-
-	err := c.Create(ctx, catalog)
-	return catalog, err
-}
